Separate filter component - pagination now works, but filtering broken
